+++ src/Root.jsx
@@ -6,10 +6,10 @@ import AdminPanel from './pages/admin/admin'
 import IndexPage from './pages/index/index'
 import GithubOAuthPage from './pages/auth/github'
 import AuthPage from './pages/auth/auth'
-import SignupPage from './pages/auth/signup'
 import SigninPage from './pages/auth/signin'
 import ClippingPage from './pages/clipping/clipping'
 import ReportYearlyPage from './pages/report/yearly'
+import AuthPhone from './pages/auth/auth-phone'
 
 const PrivacyPolicy = React.lazy(() => import(/* webpackPrefetch: true */ './pages/policy/privacy'))
 const DashContainer = React.lazy(() => import(/* webpackPrefetch: true */ './components/dashboard-container/container'))
@@ -34,8 +34,8 @@ class Root extends React.Component {
             <ReleasePage path="/release/:platform" />
             <ReportYearlyPage path='/report/yearly' />
             <AuthPage path="/auth">
-              <SigninPage path="signin" />
-              <SignupPage path="signup" />
+              <AuthPhone path='phone' />
+              <SigninPage path='signin' />
             </AuthPage>
             <GithubOAuthPage path="/oauth/github" />
             <DashContainer path="dash/:userid">