+++ src/constants/config.ts
@@ -16,3 +16,9 @@ export const SignInWithAppleOptions = {
 }
 
 export const CDN_DEFAULT_DOMAIN = 'https://clippingkk-cdn.annatarhe.com'
+
+export const LEANCLOUD = {
+  APP_ID: 'aA4dKCxL9noDX5CyYp2AdnHl-9Nh9j0Va',
+  APP_KEY: 'QQ7h7pbOY4xFWEap8fVuFtcl',
+  SERVER_URL: 'https://clippingkk-lc.annatarhe.cn'
+}